+++ backend/src/main/java/de/hs_rm/backend/messaging/response/EggcreatedResponse.java
@@ -1,28 +1,28 @@
 package de.hs_rm.backend.messaging.response;
 
 public class EggcreatedResponse extends BaseResponse {
-    private double positionX;
-    private  double positionY;
+    private int positionX;
+    private int positionY;
 
-    public EggcreatedResponse(double posX, double posY){
+    public EggcreatedResponse(int posX, int posY){
         super("itemCollected","ok");
         this.positionX = posX;
         this.positionY = posY;
     }
 
-    public double getPositionX() {
+    public int getPositionX() {
         return positionX;
     }
 
-    public void setPositionX(double positionX) {
+    public void setPositionX(int positionX) {
         this.positionX = positionX;
     }
 
-    public double getPositionY() {
+    public int getPositionY() {
         return positionY;
     }
 
-    public void setPositionY(double positionY) {
+    public void setPositionY(int positionY) {
         this.positionY = positionY;
     }
 }
